% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmAnisotropy.R
\name{anis2D.par2A}
\alias{anis2D.par2A}
\title{Produce anisotropy matrix from angle and anisotropy ratios}
\usage{
anis2D.par2A(ratio, angle)
}
\arguments{
\item{ratio}{an anisotropy ratio (min/max range)}

\item{angle}{direction of maximum range, i.e. largest spatial continuity, measured
counterclockwise from East}
}
\value{
a 3x3 matrix of anisotropy, in Cholesky form
}
\description{
Produce anisotropy matrix (in Cholesky decomposition) from angle and anisotropy ratios
}
