% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmDataFrameStack.R
\name{as.array.DataFrameStack}
\alias{as.array.DataFrameStack}
\title{Convert a stacked data frame into an array}
\usage{
\method{as.array}{DataFrameStack}(x, ...)
}
\arguments{
\item{x}{input \code{\link[=DataFrameStack]{DataFrameStack()}}}

\item{...}{generic consistency}
}
\value{
the data recasted as an array with appropriate names.
}
\description{
Recast a \code{\link[=DataFrameStack]{DataFrameStack()}} into a named array.
}
\examples{
ll = lapply(1:3, function(i) as.data.frame(matrix(1:10+(i-1)*10, ncol=2)))
dfs = DataFrameStack(ll, stackDimName="rep")
as.array(dfs)

}
