% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionsCompatibility.R
\name{as.CompLinModCoReg}
\alias{as.CompLinModCoReg}
\alias{as.CompLinModCoReg.CompLinModCoReg}
\alias{as.CompLinModCoReg.LMCAnisCompo}
\title{Recast a model to the variogram model of package "compositions"}
\usage{
as.CompLinModCoReg(v, ...)
}
\arguments{
\item{v}{variogram model object to convert}

\item{...}{further parameters for generic functionality}
}
\value{
The variogram model recast to "CompLinModCoReg"
}
\description{
Recast a variogram model specified in any of the models of "gstat" or "gmGeostats" in
the format of \code{\link[compositions:CompLinModCo]{compositions::CompLinModCoReg()}}
}
