% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variograms.R
\name{as.gmEVario}
\alias{as.gmEVario}
\alias{as.gmEVario.gstatVariogram}
\alias{as.gmEVario.logratioVariogram}
\alias{as.gmEVario.logratioVariogramAnisotropy}
\title{Convert empirical structural function to gmEVario format}
\usage{
as.gmEVario(vgemp, ...)
}
\arguments{
\item{vgemp}{variogram/covariance function to be converted}

\item{...}{further parameters}
}
\value{
the empirical covariance function or variogram, recasted to class
\code{gmEVario}. This is a generic function. Methods exist for objects of
class \code{logratioVariogram}\code{logratioVariogramAnisotropy}
(for compositional data) and \code{gstatVariogram}
(from package \code{gstat}).
}
\description{
Convert empirical covariance functions or variograms to the format gmEVario
of package gmGeostats
}
\seealso{
Other gmEVario functions: 
\code{\link{gsi.EVario2D}()},
\code{\link{gsi.EVario3D}()},
\code{\link{ndirections}()},
\code{\link{plot.gmEVario}()},
\code{\link{variogramModelPlot}()}
}
\concept{gmEVario functions}
