% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmDataFrameStack.R
\name{as.list.DataFrameStack}
\alias{as.list.DataFrameStack}
\title{Convert a stacked data frame into a list of data.frames}
\usage{
\method{as.list}{DataFrameStack}(x, ...)
}
\arguments{
\item{x}{input \code{\link[=DataFrameStack]{DataFrameStack()}}}

\item{...}{generic consistency}
}
\value{
the data recasted as list of data.frames
}
\description{
Recast a \code{\link[=DataFrameStack]{DataFrameStack()}} into a list of data.frames
}
\examples{
ar = array(1:30, dim = c(5,2,3), dimnames=list(obs=1:5, vars=c("A","B"), rep=1:3))
dfs = DataFrameStack(ar, stackDim="rep")
as.list(dfs)
}
