% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionsCompatibility.R
\name{as.logratioVariogram}
\alias{as.logratioVariogram}
\alias{as.logratioVariogram.logratioVariogram}
\alias{as.logratioVariogram.gmEVario}
\title{Recast empirical variogram to format logratioVariogram}
\usage{
as.logratioVariogram(vgemp, ...)
}
\arguments{
\item{vgemp}{empirical variogram}

\item{...}{parameters for generic functionality}
}
\value{
the same model in the new format.
}
\description{
Recast an empirical compositional variogram of any sort to a variation-variogram of class
"logratioVariogram".
}
