% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionsCompatibility.R
\name{as.logratioVariogramAnisotropy}
\alias{as.logratioVariogramAnisotropy}
\alias{as.logratioVariogramAnisotropy.default}
\alias{as.logratioVariogramAnisotropy.logratioVariogram}
\alias{as.logratioVariogramAnisotropy.logratioVariogramAnisotropy}
\title{Convert empirical variogram to "logratioVariogramAnisotropy"}
\usage{
as.logratioVariogramAnisotropy(vgemp, ...)

\method{as.logratioVariogramAnisotropy}{default}(vgemp, ...)

\method{as.logratioVariogramAnisotropy}{logratioVariogram}(vgemp, ...)

\method{as.logratioVariogramAnisotropy}{logratioVariogramAnisotropy}(vgemp, ...)
}
\arguments{
\item{vgemp}{an empirical variogram}

\item{...}{further parameters}
}
\value{
The empirical variogram as a  "logratioVariogramAnisotropy" object
}
\description{
Convert an empirical variogram from any format to class "logratioVariogramAnisotropy"
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method, making use of \code{as.logratioVariogram()}

\item \code{logratioVariogram}: method for "logratioVariogram" class

\item \code{logratioVariogramAnisotropy}: identity transformation
}}

