% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmDataFrameStack.R
\name{dimnames.DataFrameStack}
\alias{dimnames.DataFrameStack}
\alias{dimnames,Spatial-method}
\title{Return the dimnames of a DataFrameStack}
\usage{
\method{dimnames}{DataFrameStack}(x)

\S4method{dimnames}{Spatial}(x)
}
\arguments{
\item{x}{a \code{\link[=DataFrameStack]{DataFrameStack()}} object}
}
\value{
a list (possibly named) with the element names of each of the three dimensions
}
\description{
Return the dimnames of a \code{\link[=DataFrameStack]{DataFrameStack()}}, i.e. the three dimensions
}
\section{Functions}{
\itemize{
\item \code{dimnames,Spatial-method}: dimnames method for all Spatial*DataFrame objects of
package \code{sp} which data slot contains a \code{\link[=DataFrameStack]{DataFrameStack()}}
}}

