% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmGaussianMethodParameters-class}
\alias{gmGaussianMethodParameters-class}
\title{parameters for Spatial Gaussian methods of any kind}
\description{
abstract class, containing any parameter specification for a spatial algorithm
for interpolation, simulation or validation making use of Gaussian assumptions
}
