% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmGaussianSimulationAlgorithm-class}
\alias{gmGaussianSimulationAlgorithm-class}
\title{parameters for Gaussian Simulation methods}
\description{
abstract class, containing any parameter specification of a spatial simulation algorithm
exploiting a Gaussian two-point model structure
}
