% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmMPSParameters-class}
\alias{gmMPSParameters-class}
\title{parameters for Multiple-Point Statistics methods}
\description{
abstract class, containing any parameter specification of a spatial multipoint algorithm
}
