% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmNeighbourhoodSpecification-class}
\alias{gmNeighbourhoodSpecification-class}
\title{Neighbourhood description}
\description{
abstract class, containing any specification of a spatial neighbourhood
}
