% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmSimulationAlgorithm-class}
\alias{gmSimulationAlgorithm-class}
\title{Parameter specification for a spatial simulation algorithm}
\description{
abstract class, containing any parameter specification for a spatial simulation algorithm
}
