% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{class}
\name{gmSpatialDataContainer-class}
\alias{gmSpatialDataContainer-class}
\title{General description of a spatial data container}
\description{
abstract class, containing any specification of a spatial data container
}
\details{
setClassUnion(name="gmTrainingImage",
members=c("SpatialGridDataFrame",
"SpatialPixelsDataFrame")
)
}
