% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmSpatialMethodParameters-class}
\alias{gmSpatialMethodParameters-class}
\title{Parameter specification for any spatial method}
\description{
abstract class, containing any parameter specification for
any spatial method. Members of this class are \linkS4class{gmNeighbourhoodSpecification}
\linkS4class{gmMPSParameters} and \linkS4class{gmValidationStrategy}.
}
