% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmUnconditionalSpatialModel-class}
\alias{gmUnconditionalSpatialModel-class}
\title{General description of a spatial model}
\description{
abstract class, containing any specification of an unconditional
spatial model
}
