% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmValidationStrategy-class}
\alias{gmValidationStrategy-class}
\title{Validation strategy description}
\description{
abstract class, containing any specification of a validation strategy for spatial models
}
