% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{gsi.calcCgram}
\alias{gsi.calcCgram}
\title{Compute covariance matrix oout of locations}
\usage{
gsi.calcCgram(X, Y, vgram, ijEqual = FALSE)
}
\arguments{
\item{X}{matrix, coordinates of the first set of locations}

\item{Y}{matrix, coordinates of the first set of locations (often Y=X, and ijEqual=TR)}

\item{vgram}{covariogram model, of format "gmCgram"}

\item{ijEqual}{logical, are X==Y? if you put TRUE, the function will only compute half of
the points, so be sure that this is right!}
}
\value{
a matrix of elements of the covariance, in NxM blocks of DxD elements
(being N and M resp. the nr. of rows of X and Y, and D the nr of variables)
}
\description{
internal function to compute the variogram model for all combinations of two
}
