% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionsCompatibility.R
\name{gsi.orig}
\alias{gsi.orig}
\alias{gsi.getV}
\title{extract information about the original data, if available}
\usage{
gsi.orig(x, y = NULL)
}
\arguments{
\item{x}{an rmult object}

\item{y}{possibly another rmult object}
}
\value{
The original untransformed data (with a compositional class), resp
the TRANSPOSED INVERSE matrix, i.e. the one to recover the original components.
}
\description{
originally implemented in package:compositions
}
