% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{gsi.TurningBands}
\alias{gsi.TurningBands}
\title{Internal function, unconditional turning bands realisations}
\usage{
gsi.TurningBands(X, vgram, nBands, nsim = NULL)
}
\arguments{
\item{X}{matrix of coordinates of locations where realisations are desired}

\item{vgram}{covariogram model, of format "gmCgram"}

\item{nBands}{number of bands to use}

\item{nsim}{number of realisations to return}
}
\value{
an array with (npoint, nvar, nsim)-elements, being npoint=nrow(X)
and nvar = nr of variables in vgram
}
\description{
Internal function to compute unconditional turning bands simulations
}
