% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratools.R
\name{has.missings.data.frame}
\alias{has.missings.data.frame}
\title{Check presence of missings
check presence of missings in a data.frame}
\usage{
\method{has.missings}{data.frame}(x, mc = NULL, ...)
}
\arguments{
\item{x}{data, of class data.frame}

\item{mc}{not used}

\item{...}{not used}
}
\value{
A single true/false response about the presence of any missing value
on the whole data set
}
\description{
Check presence of missings
check presence of missings in a data.frame
}
\examples{
library(compositions)
data(Windarling)
has.missings(Windarling)
head(Windarling)
Windarling[1,1] = NA
head(Windarling)
has.missings(Windarling)
}
