% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{image.mask}
\alias{image.mask}
\title{Image method for mask objects}
\usage{
\method{image}{mask}(x, col = c(NA, 2), ...)
}
\arguments{
\item{x}{a mask}

\item{col}{a two-color vector for the color (oustide, inside) the mask}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Plot a mask of a 2D grid; see \code{\link[=constructMask]{constructMask()}} for an example
}
