% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variograms.R
\name{is.anisotropySpecification}
\alias{is.anisotropySpecification}
\title{Check for any anisotropy class}
\usage{
is.anisotropySpecification(x)
}
\arguments{
\item{x}{object to check}
}
\value{
a logical, TRUE if the object is an anisotropy specification; FALSE otherwise
}
\description{
Check that an object contains a valid specification of anisotropy, in any form
}
\examples{
a =  anis2D_par2A(0.5, 30)
a
is.anisotropySpecification(a)
}
\seealso{
Other anisotropy: 
\code{\link{AnisotropyRangeMatrix}()},
\code{\link{AnisotropyScaling}()},
\code{\link{anis_GSLIBpar2A}()},
\code{\link{as.AnisotropyRangeMatrix}()},
\code{\link{as.AnisotropyScaling}()}
}
\concept{anisotropy}
