% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variograms.R
\name{is.isotropic}
\alias{is.isotropic}
\title{Check for anisotropy of a theoretical variogram}
\usage{
is.isotropic(x, tol = 1e-10, ...)
}
\arguments{
\item{x}{variogram or covariance model object}

\item{tol}{tolerance for}

\item{...}{extra arguments for generic  functionality}
}
\value{
Generic function. Returns of boolean answering the question of the name,
or NA if object \code{x} does not contain a known theoretical variogram
}
\description{
Checks for anisotropy of a theoretical variogram or covariance function model
}
