% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmValidationStrategy.R
\name{LeaveOneOut}
\alias{LeaveOneOut}
\title{Specify the leave-one-out strategy for validation of a spatial model}
\usage{
LeaveOneOut()
}
\value{
an object of class c("LeaveOneOut", "NfoldCrossValidation") to be used
in a call to \code{\link[=validate]{validate()}}
}
\description{
Function to specify the leave-one-out strategy for validation of a spatial model
}
\examples{
LeaveOneOut()
}
\seealso{
Other validation functions: 
\code{\link{NfoldCrossValidation}},
\code{\link{validate}()}
}
\concept{validation functions}
