% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{mean.accuracy}
\alias{mean.accuracy}
\title{Mean accuracy}
\usage{
\method{mean}{accuracy}(x, ...)
}
\arguments{
\item{x}{accuracy curve obtained with \code{\link[=accuracy]{accuracy()}}}

\item{...}{generic functionality, not used}
}
\value{
The value of the mean accuracy after Deutsch (1997); details
can be found in \code{\link[=precision.accuracy]{precision()}}.
}
\description{
Mean method for accuracy curves, after Deutsch (1997)
}
\seealso{
Other accuracy functions: 
\code{\link{accuracy}()},
\code{\link{plot.accuracy}()},
\code{\link{precision}()},
\code{\link{validate}()},
\code{\link{xvErrorMeasures.default}()},
\code{\link{xvErrorMeasures}()}
}
\concept{accuracy functions}
