% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncorrelationTest.R
\name{mean.spatialDecorrelationMeasure}
\alias{mean.spatialDecorrelationMeasure}
\title{Average measures of spatial decorrelation}
\usage{
\method{mean}{spatialDecorrelationMeasure}(x, ...)
}
\arguments{
\item{x}{the outcome of a call to \code{\link[=spatialDecorrelation]{spatialDecorrelation()}}}

\item{...}{further arguments to mean}
}
\value{
a mean for each measure available on \code{x}, i.e. this can be a vector.
The output is named.
}
\description{
Compute average measres of spatial decorrelation out of the result of a call
to \code{\link[=spatialDecorrelation]{spatialDecorrelation()}}
}
\seealso{
\code{\link[=spatialDecorrelation]{spatialDecorrelation()}} for an example
}
