% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variograms.R
\docType{class}
\name{ModelStructuralFunctionSpecification-class}
\alias{ModelStructuralFunctionSpecification-class}
\title{Structural function model specification}
\description{
Abstract class, containing any specification of a variogram (or covariance) model
}
