% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variograms.R
\name{+.gmCgram}
\alias{+.gmCgram}
\title{Combination of gmCgram variogram structures}
\usage{
\method{+}{gmCgram}(x, y)
}
\arguments{
\item{x}{\code{gmCgram} variogram object}

\item{y}{\code{gmCgram} variogram object}
}
\value{
The combined nested structures
}
\description{
combination of nested structures of a gmCgram object
}
\examples{
utils::data("variogramModels")
v1 = setCgram(type=vg.Gau, sill=diag(2), anisRanges = 3*diag(c(3,1)))
v2 = setCgram(type=vg.Exp, sill=0.3*diag(2), anisRanges = 0.5*diag(2))
vm = v1+v2
}
