% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{print.mask}
\alias{print.mask}
\title{Print method for mask objects}
\usage{
\method{print}{mask}(x, ...)
}
\arguments{
\item{x}{mask to print}

\item{...}{ignored}
}
\value{
the summary of number of nodes inside/outside the mask
}
\description{
Print method for mask objects. See \code{\link[=constructMask]{constructMask()}} for examples.
If you want to see the whole content of the mask, then use \code{unclass(...)}
}
\seealso{
Other masking functions: 
\code{\link{constructMask}()},
\code{\link{getMask}()},
\code{\link{setMask}()},
\code{\link{unmask}()}
}
\concept{masking functions}
