% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratools.R
\name{spectralcolors}
\alias{spectralcolors}
\title{Spectral colors palette
based on the RColorBrewer::brewer.pal(11,"Spectral")}
\usage{
spectralcolors(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
a palette, i.e. a list of colors, from dark blue to dark red over pale yellow.
}
\description{
Spectral colors palette
based on the RColorBrewer::brewer.pal(11,"Spectral")
}
\examples{
(cls=spectralcolors(20))
}
