% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmDataFrameStack.R
\name{stackDim,Spatial-method}
\alias{stackDim,Spatial-method}
\title{Get name/index of the stacking dimension of a Spatial object}
\usage{
\S4method{stackDim}{Spatial}(x)
}
\arguments{
\item{x}{a \code{Spatial} object which \code{data} slot is a \code{DataFrameStack}}
}
\value{
see \code{\link[=stackDim]{stackDim()}} for details
}
\description{
Get name/index of the stacking dimension of a Spatial object which
data slot is of class \code{\link[=DataFrameStack]{DataFrameStack()}}
}
\seealso{
\code{\link[=stackDim]{stackDim()}}
}
