% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparations.R
\name{variogram}
\alias{variogram}
\title{Empirical variogram or covariance function}
\usage{
variogram(object, ...)
}
\arguments{
\item{object}{spatial data container, with special methods according to the class}

\item{...}{further parameters for generic functionality}
}
\value{
An empirical variogram for the provided data.
}
\description{
Generic function to compute the empirical variogram or covariance function from actual data
}
\seealso{
Other gmEVario functions: 
\code{\link{as.gmEVario.gstatVariogram}()},
\code{\link{gsi.EVario2D}()},
\code{\link{gsi.EVario3D}()},
\code{\link{ndirections}()},
\code{\link{plot.gmEVario}()},
\code{\link{variogramModelPlot}()}
}
\concept{gmEVario functions}
