% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionsCompatibility.R
\name{variogramModelPlot.logratioVariogram}
\alias{variogramModelPlot.logratioVariogram}
\title{Quick plotting of empirical and theoretical logratio variograms
Quick and dirty plotting of empirical logratio variograms with or without their models}
\usage{
\method{variogramModelPlot}{logratioVariogram}(vg, model = NULL, col = rev(rainbow(ndirections(vg))), ...)
}
\arguments{
\item{vg}{empirical variogram or covariance function}

\item{model}{optional, theoretical variogram or covariance function}

\item{col}{colors to use for the several directional variograms}

\item{...}{further parameters to \code{plot.logratioVariogramAnisotropy()}}
}
\value{
The function is primarily called for producing a plot. However, it
invisibly returns the graphical parameters active before the call
occurred. This is useful for constructing complex diagrams, by adding layers
of info. If you want to "freeze" your plot, embed your call in another
call to \code{\link{par}}, e.g. \code{par(variogramModelPlot(...))}.
}
\description{
Quick plotting of empirical and theoretical logratio variograms
Quick and dirty plotting of empirical logratio variograms with or without their models
}
\seealso{
Other variogramModelPlot: 
\code{\link{variogramModelPlot.gstatVariogram}()},
\code{\link{variogramModelPlot}()}
}
\concept{variogramModelPlot}
