% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Windarling}
\alias{Windarling}
\title{Ore composition of a bench at a mine in Windarling, West Australia.}
\format{
A data.frame with 1600 rows and 16 columns:
\describe{
\item{Hole_id}{ID of the observation}
\item{Sample.West}{indicator for belonging to a subsample on the West sector}
\item{Sample.East}{indicator for belonging to a subsample on the East sector}
\item{West}{indicator for belonging to the Western wing of the bench}
\item{East}{indicator for belonging to the Eastern wing of the bench}
\item{Easting}{Easting (X) coordinate of the sample}
\item{Northing}{Northing (Y) coordinate of the sample}
\item{Lithotype}{factor, indicating material type of the sample, one of: basalt, goethite, magnetite, schist}
\item{Fe}{percent of Iron in the sample}
\item{P}{percent of Phosphorus in the sample}
\item{SiO2}{percent of Silica in the sample}
\item{Al2O3}{percent of Alumina in the sample}
\item{S}{percent of Sulphur in the sample}
\item{Mn}{percent of Magnanese in the sample}
\item{CL}{percent of Chlorine in the sample}
\item{LOI}{percent Loss-on-Ignition of the sample}
}
}
\source{
Ward (2015) Compositions, logratios and
geostatistics: An application to iron ore. MSc Thesis
Edith Cowan University; \url{https://ro.ecu.edu.au/theses/1581/}
}
\usage{
Windarling
}
\description{
A dataset containing the geochemical composition (and some extra variables)
of a grade control study of a mine bench at Windarling, West Australia.
}
\section{License}{
  CC BY-SA 4.0
}

\keyword{datasets}
