% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{xvErrorMeasures.default}
\alias{xvErrorMeasures.default}
\title{Cross-validation errror measures}
\usage{
\method{xvErrorMeasures}{default}(x, krigVar, observed, output = "MSDR1", ...)
}
\arguments{
\item{x}{a vector containing the predicted values}

\item{krigVar}{a vector containing the kriging variances}

\item{observed}{a vector containing the true values}

\item{output}{which output do you want? a vector of one or several of  c("ME","MSE","MSDR","Mahalanobis")}

\item{...}{extra arguments for generic functionality}
}
\description{
Compute one or more error measures from cross-validation output
}
\seealso{
Other accuracy functions: 
\code{\link{accuracy}()},
\code{\link{mean.accuracy}()},
\code{\link{plot.accuracy}()},
\code{\link{precision}()},
\code{\link{validate}()},
\code{\link{xvErrorMeasures}()}
}
\concept{accuracy functions}
