% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variograms.R
\docType{class}
\name{EmpiricalStructuralFunctionSpecification-class}
\alias{EmpiricalStructuralFunctionSpecification-class}
\title{Empirical structural function specification}
\description{
Abstract class, containing any specification of an empirical variogram
(or covariance function, or variations). Members must implement a coercion method to
class "gmEVario" (see \code{\link[=gsi.EVario2D]{gsi.EVario2D()}} for an example), and (possibly) coercion to
class "gstatVariogram" (see \code{\link[gstat:variogram]{gstat::variogram()}})
}
