% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractClasses.R
\docType{class}
\name{gmTrainingImage-class}
\alias{gmTrainingImage-class}
\title{MPS training image class}
\description{
abstract class, containing any specification of a multiple-point
training image. This must be analogous to \code{\link[sp:SpatialGridDataFrame]{sp::SpatialGridDataFrame()}}, and
implement a method for \code{\link[sp:SpatialGrid]{sp::getGridTopology()}} and to coerce it to
\link[sp:SpatialGridDataFrame-class]{sp::SpatialGridDataFrame}.
}
